refactor: suggestions
